pipes/api-weather-raw/requirements.txt
requests==2.32.3
google-cloud-bigquery==3.25.0
Flask==3.0.3